Refactor installation dependency chain
I conferred with @marcoceppi this morning about the charm-tools and
charm dependency chain and the future facing move to snap delivery.

We're currently pressed as we cant deliver software via snaps inside
docker unless we're running systemd amenities. Docker doesn't play well
with running an init, so we're at an impass. Here's the current
packaging notes:

 - Instal charm-tools from pypi. This will always be up to date with the
   latest stable release of charm-tools
 - For now, install the charm command from apt, until we have the
   delivery pipeline sorted
 - any derivative works should actively try not to unpin versioned
   dependencies such as pyyaml. So a simple `pip install` vs `pip
   install -U` will suffice.

This build was tested as a base for cwrbox and resolves the dependency
chain issues that surfaced on 8/30/2016