Reconcile with jujubox
This cleans up the existing charmbox codebase in an attempt to bring it
in alignment with whats happening with jujubox. I've removed a lot of
the excess configuration, and user-munging.

I had to re-introduce the ARG defined in jujubox, as docker does not
seem to port this between images. Unfortunate I know.